Revert changes to Compiler_PlatformWindows.cpp
