+++ Aurora/cmake/ProjectFiles.cmake
@@ -13,7 +13,6 @@ aux_source_directory(RuntimeCompiler RuntimeCompiler_SRCS)
 aux_source_directory(RuntimeCompiler/SimpleFileWatcher SimpleFileWatcher_SRCS)
 
 if(UNIX)
-    list(REMOVE_ITEM RuntimeCompiler_SRCS "RuntimeCompiler/Compiler_PlatformVS.cpp")
     list(REMOVE_ITEM RuntimeCompiler_SRCS "RuntimeCompiler/Compiler_PlatformMinGW.cpp")
 	list(REMOVE_ITEM RuntimeCompiler_SRCS "RuntimeCompiler/Compiler_PlatformWindows.cpp")
 	list(REMOVE_ITEM SimpleFileWatcher_SRCS "RuntimeCompiler/SimpleFileWatcher/FileWatcherWin32.cpp")
@@ -24,7 +23,7 @@ if(UNIX)
 	endif()
 else()
     if(MINGW_MAKEFILE)
-        list(REMOVE_ITEM RuntimeCompiler_SRCS "RuntimeCompiler/Compiler_PlatformVS.cpp")
+        list(REMOVE_ITEM RuntimeCompiler_SRCS "RuntimeCompiler/Compiler_PlatformWindows.cpp")
     else()
         list(REMOVE_ITEM RuntimeCompiler_SRCS "RuntimeCompiler/Compiler_PlatformMinGW.cpp")
     endif()
